aider: Created the `Library` class which is a child class of `GeoEntity` and handles CRUD operations for data management.
# Aider chat conversation:

USER: so do we now move to creating the other geoentities in this sprint?

ASSISTANT: Yes, we can proceed with creating the other geographical entities. However, it would be helpful to know which other geographical entities you are considering for this sprint. This will allow us to plan the creation of these classes effectively.

USER: well I suppose it's just the library then. And given that the library handles all the CRUD operations for data management, it will probably need to import some tools for accessing and working with various database types as described in the sprint 1 discovery.

ASSISTANT: Alright, we can proceed with creating the `Library` class. This class will be responsible for handling all the CRUD (Create, Read, Update, Delete) operations for data management. As you mentioned, it will need to import some tools for accessing and working with various database types.

Here are the steps we need to follow:

1. Create a new file for the `Library` class. This file will be located in the `src/hamlet/library` directory. Let's name the file `library.py`.

2. In the `library.py` file, we will define the `Library` class. This class will also be a child class of `GeoEntity`, similar to the `Towncenter` and `Hamlet` classes. It will have its own attributes and methods specific to a library.

3. We will need to import the necessary tools for accessing and working with various database types. These tools will be used in the methods of the `Library` class to perform the CRUD operations.

Here is the proposed code for the `Library` class:

```python
src/hamlet/library/library.py
<<<<<<< HEAD
=======
from ..hamlet import GeoEntity

class Library(GeoEntity):
    def __init__(self, name, location, database=None):
        super().__init__(name, location)
        self.database = database if database is not None else {}

    def create_data(self, data):
        # code to create data in the database

    def read_data(self, data_id):
        # code to read data from the database

    def update_data(self, data_id, new_data):
        # code to update data in the database

    def delete_data(self, data_id):
        # code to delete data from the database
>>>>>>> updated
```

Please note that the methods `create_data`, `read_data`, `update_data`, and `delete_data` are placeholders. The actual code for these methods will depend on the specific database tools that we will use.