aider: Rename the functions in the `Library` class to reflect that they are SQL functions.
# Aider chat conversation:

USER: rename the the functions to reflect that they are sql functions.  We will definitely be including knowledge graphs but those will be populated from relational database data anyway.  So really just recording all info with sql is the main priority

ASSISTANT: Alright, we can rename the functions in the `Library` class to reflect that they are SQL functions. Here are the changes we need to make:

1. In the `src/hamlet/library/library.py` file, we will rename the `create_data`, `read_data`, `update_data`, and `delete_data` methods to `create_data_sql`, `read_data_sql`, `update_data_sql`, and `delete_data_sql` respectively.

Here is the proposed code change:

```python
src/hamlet/library/library.py
<<<<<<< HEAD
    def create_data(self, data):
        # code to create data in the database

    def read_data(self, data_id):
        # code to read data from the database

    def update_data(self, data_id, new_data):
        # code to update data in the database

    def delete_data(self, data_id):
        # code to delete data from the database
=======
    def create_data_sql(self, data):
        # code to create data in the SQL database

    def read_data_sql(self, data_id):
        # code to read data from the SQL database

    def update_data_sql(self, data_id, new_data):
        # code to update data in the SQL database

    def delete_data_sql(self, data_id):
        # code to delete data from the SQL database
>>>>>>> updated
```

Please note that the actual code for these methods will depend on the specific SQL database tools that we will use.